- move _visual_compiler_last from detect.py into win.py, renaming to the latest_visual_studio_version_installed
